requirements.txt
asgiref==3.8.1
attrs==22.1.0
beautifulsoup4==4.11.1
bidict==0.22.0
coverage==6.4.4
crispy-bootstrap4==2022.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==5.2.5
django-admin-autocomplete-filter==0.7.1
django-admin-list-filter-dropdown==1.0.3
django-rangefilter>=0.9.0
django-admin-searchable-dropdown==1.2
django-bootstrap-datepicker-plus==5.0.5
django-crispy-forms==2.0
django-durationwidget2==1.0.6
django-extensions==3.2.1
django-import-export==3.0.2
django-revproxy==0.10.0
et-xmlfile==1.1.0
execnet==1.9.0
fdfgen==0.16.1
freezegun==1.2.2
icalendar==5.0.3
iniconfig==1.1.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
PyPDF2==2.10.3
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-env==0.6.2
pytest-forked==1.4.0
pytest-xdist==2.5.0
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0.1
qrcode==7.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.3.0
tomli==2.0.1
urllib3==1.26.12
waitress==2.1.2
WebOb==1.8.7
WebTest==3.0.0
WSGIProxy2==0.5.1
xlrd==2.0.1
xlwt==1.3.0
zope.cachedescriptors==4.3.1
zope.event==4.5.0
zope.interface==5.4.0
zope.schema==6.2.0
zope.testbrowser==5.6.1
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-filter==23.5
pymediainfo==6.1.0
